+++ services/federator/src/Federator/Options.hs
@@ -62,9 +62,11 @@ data RunSettings = RunSettings
   { -- | Would you like to federate with everyone or only with a select set of other wire-server installations?
     federationStrategy :: FederationStrategy,
     useSystemCAStore :: Bool,
-    remoteCAStore :: Maybe FilePath
+    remoteCAStore :: Maybe FilePath,
+    clientCertificate :: FilePath,
+    clientPrivateKey :: FilePath
   }
-  deriving (Show, Generic)
+  deriving (Eq, Show, Generic)
 
 instance FromJSON RunSettings